Added option for absorption of lower tiered damage
For example, whenyou get three points damage of a lower tier and one point damage of a higher tier, you end
up with the one point damage of higher tier and two points damage of lower tier when the optio is set. The
one point of lower tiered damage that seems missing got absorbed by the one point damage of higher tier. Some
systems work with such a damage notation. This toggle is a convenience option so that you do not have to do
the (fairly simple) computations in your head.